+++ packages/react-sdk/src/components/Video/VideoPreview.tsx
@@ -7,8 +7,13 @@ import {
 } from 'react';
 import clsx from 'clsx';
 import { disposeOfMediaStream } from '@stream-io/video-client';
-import { BaseVideo } from '../../core/components/Video/BaseVideo';
-import { DEVICE_STATE, useMediaDevices } from '../../core/contexts';
+import { BaseVideo } from '../../core/components/Video';
+import {
+  DEVICE_STATE,
+  useMediaDevices,
+  useOnUnavailableVideoDevices,
+  useVideoDevices,
+} from '../../core';
 import { LoadingIndicator } from '../LoadingIndicator';
 
 const DefaultDisabledVideoPreview = () => {
@@ -47,19 +52,24 @@ export const VideoPreview = ({
   VideoErrorPreview = DefaultVideoErrorPreview,
 }: VideoPreviewProps) => {
   const [stream, setStream] = useState<MediaStream>();
-
   const {
-    videoDevices,
     selectedVideoDeviceId,
     getVideoStream,
     initialVideoState,
     setInitialVideoState,
   } = useMediaDevices();
+  // When there are 0 video devices (e.g. when laptop lid closed),
+  // we do not restart the video automatically when the device is again available,
+  // but rather leave turning the video on manually to the user.
+  useOnUnavailableVideoDevices(() =>
+    setInitialVideoState(DEVICE_STATE.stopped),
+  );
+  const videoDevices = useVideoDevices();
 
   useEffect(() => {
-    if (!initialVideoState.enabled || videoDevices.length === 0) return;
+    if (!initialVideoState.enabled) return;
 
-    getVideoStream(selectedVideoDeviceId)
+    getVideoStream({ deviceId: selectedVideoDeviceId })
       .then((s) => {
         setStream((previousStream) => {
           if (previousStream) {