+++ packages/styling/CHANGELOG.md
@@ -2,6 +2,83 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.0.1-alpha.30](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.29...@stream-io/video-styling-0.0.1-alpha.30) (2023-05-18)
+
+
+
+## [0.0.1-alpha.29](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.28...@stream-io/video-styling-0.0.1-alpha.29) (2023-05-18)
+
+
+
+## [0.0.1-alpha.28](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.27...@stream-io/video-styling-0.0.1-alpha.28) (2023-05-18)
+
+
+
+## [0.0.1-alpha.27](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.26...@stream-io/video-styling-0.0.1-alpha.27) (2023-05-17)
+
+
+
+## [0.0.1-alpha.26](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.25...@stream-io/video-styling-0.0.1-alpha.26) (2023-05-17)
+
+
+
+## [0.0.1-alpha.25](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.24...@stream-io/video-styling-0.0.1-alpha.25) (2023-05-17)
+
+
+
+## [0.0.1-alpha.24](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.23...@stream-io/video-styling-0.0.1-alpha.24) (2023-05-17)
+
+
+
+## [0.0.1-alpha.23](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.22...@stream-io/video-styling-0.0.1-alpha.23) (2023-05-17)
+
+
+
+## [0.0.1-alpha.22](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.21...@stream-io/video-styling-0.0.1-alpha.22) (2023-05-17)
+
+
+
+## [0.0.1-alpha.21](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.20...@stream-io/video-styling-0.0.1-alpha.21) (2023-05-16)
+
+
+
+## [0.0.1-alpha.20](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.19...@stream-io/video-styling-0.0.1-alpha.20) (2023-05-16)
+
+
+
+## [0.0.1-alpha.19](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.18...@stream-io/video-styling-0.0.1-alpha.19) (2023-05-16)
+
+
+
+## [0.0.1-alpha.18](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.17...@stream-io/video-styling-0.0.1-alpha.18) (2023-05-16)
+
+
+### Bug Fixes
+
+* **styling:** set min-height to ParticipantView in the legacy grid ([#506](https://github.com/GetStream/stream-video-js/issues/506)) ([696608b](https://github.com/GetStream/stream-video-js/commit/696608b73256135fb34abe8e6cf19d04a309671e))
+
+
+
+## [0.0.1-alpha.17](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.16...@stream-io/video-styling-0.0.1-alpha.17) (2023-05-16)
+
+
+
+## [0.0.1-alpha.16](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.15...@stream-io/video-styling-0.0.1-alpha.16) (2023-05-16)
+
+
+
+## [0.0.1-alpha.15](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.14...@stream-io/video-styling-0.0.1-alpha.15) (2023-05-16)
+
+
+
+## [0.0.1-alpha.14](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.13...@stream-io/video-styling-0.0.1-alpha.14) (2023-05-16)
+
+
+
+## [0.0.1-alpha.13](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.12...@stream-io/video-styling-0.0.1-alpha.13) (2023-05-16)
+
+
+
 ## [0.0.1-alpha.12](https://github.com/GetStream/stream-video-js/compare/@stream-io/video-styling-0.0.1-alpha.11...@stream-io/video-styling-0.0.1-alpha.12) (2023-05-15)